requirements.txt
fastapi[standard]==0.113.0
pydantic==2.8.0
PyJWT==2.10.1
